Add files via upload
This Python script is an image captioning tool which generates a very much detailed captions for uploaded images using Vision Transformer (ViT) and GPT-2. It has some unique features
Pre-trained AI Model | User-friendly Interface |Automated Image Processing |Beam Search for Caption Generation | Matplotlib Integration |Loop for Continuous Usage 
How It Works:
The script loads the VisionEncoderDecoderModel (ViT-GPT2) for image captioning.
It waits for the user to upload an image or exit the loop.
Once a picture is uploaded, it:
Reads and preprocesses the picture.
Generates a caption using beam search.
Displays the image with the generated caption.
The process repeats until the user chooses to exit or stop.